Fixed sporadic xrun bug in shared memory code. This one was hard to track down. The symptom was that a few times, sometimes several hours apart, the native plugin code would take much longer than normal to complete, causing xruns for the plugin host.
It turns out that since the shared memory being used is backed by a
file, the Linux kernel likes to flush the contents in memory back to
the file occasionally. While it does this *it locks the memory
pages*, meaning that the plugin, which tries to write to the memory,
is suspended until the file write is complete.

Of course, waiting for file I/O inside the realtime code is
unacceptable, so we need shared memory that does not flush to disk.
The fix was to switch to POSIX shared memory objects, using
shm_open(), instead of normal file open().

Also fixed a trivial bug where the shared memory was opened twice
(mkstemp already opens the file, but we still opened it again
afterwards).